+++ ezcharts/__init__.py
@@ -1,6 +1,6 @@
 """Simple eCharts API."""
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 
 import argparse
 import importlib