+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Changed
 - All plots now have a toolbox by default.
+- Added ecStat 1.2.0 to vendor resources
+- Adding paging options to DataTable
 
 ## [v0.2.2]
 ### Added