+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 ### Changed
 - All plots now have a toolbox by default.
+### Added
+- CLI now configures logging.
 
 ## [v0.2.2]
 ### Added